+++ apps/cold-platform-openai/src/schemas/extraction_schemas/certificate/default.schema.ts
@@ -1,6 +1,4 @@
 import z from 'zod';
 import { baseCertificateSchema } from './base.certificate.schema';
 
-export const defaultCertificateSchema = baseCertificateSchema.extend({
-	name: z.string().describe('If the document type is a certificate then look for certificate name and place it here if found.  If not found, leave blank'),
-});
+export const defaultCertificateSchema = baseCertificateSchema;